feat: add main HTML template and remove unused GraphQL mock configurations from hps-evolve.config.ts
